requirements.txt
photodedup
python-dateutil
jrnl
flask
gunicorn